+++ DX11Starter/ParticleEmitter.h
@@ -19,7 +19,7 @@ struct ParticleEmitterDescription {
 	unsigned int maxParticles = 1000; //set once when emitter is created, unchangeable thereafter
 	float emissionStartRadius = 0.001f;
 	float emissionEndRadius = 1.0f;
-	float emissionRate = 10.0f; //particles per second
+	double emissionRate = 10.0; //particles per second
 
 	float particleMinLifetime = 10.0f; //minimum lifetime of emitted particles in seconds
 	float particleMaxLifetime = 15.0f; //maximum lifetime of emitted particles in seconds
@@ -85,7 +85,7 @@ public:
 	void SetEmissionRadii(float start, float end);
 
 	//Emission rate in particles per second
-	void SetEmissionRate(float emissionRate);
+	void SetEmissionRate(double emissionRate);
 	void SetParticleLifetime(float min, float max);
 	void SetParticleInitialScale(float min, float max);
 	void SetParticleInitialAngularVelocity(float min, float max);
@@ -123,7 +123,7 @@ protected:
 	float lifetime = 0.0f;
 	float maxLifetime = 0.0f;
 
-	float emitTimeCounter = 0.0f;
+	double emitTimeCounter = 0.0;
 
 	XMFLOAT4X4 worldMatrix = MATRIX_IDENTITY;
 	XMFLOAT3 position = ZERO_VECTOR3;
@@ -138,7 +138,7 @@ protected:
 	float emissionAngleDegrees = 0.0f;
 	float emissionAngleRadians = 0.0f;
 	unsigned int maxParticles = 0;
-	float emissionRate = 0.0f;
+	double emissionRate = 0.0;
 	float particleMinLifetime = 0.0f;
 	float particleMaxLifetime = 0.0f;
 	float particleAvgLifetime = 0.0f;
